BF-520 WriteOneTable function in DatabaseExport did not support integers(only int32). New CRFVariables table has integer fields which caused a crash.
Modified the WriteOneTable to use the actual data type of the datareader (which contains the results from the SQL Select statement) versus the former use of a delimited string that had text values containing the names of the field types.

Added INT64 data type to WriteOneTable function.

Now appears to write previously existing tables (still need to add write for new tables).


Former-commit-id: 2232746809b1725a70ba8b51bcadb4b11920bade