Further Improvements to Layer Handling
Admin layers don't get deleted and readded now. Tested and debugged new admin functionality and symbology with with baseline and control and delta data. In all cases, modified so that layers add to the map when double clicked in the tree view, rather than rebuilding the map each time. Still not getting any results for HIF.