Grid Definition Preview Map Speed and behavior improvements
Preview map syncs with the color button.
Reduced drawing time for preview maps by 10x. The CMAQ 12km went from 60 seconds to about 6 seconds. Drawing itself is 4 seconds but the row-col searches were killing it so I optimized by not opening the file and closing it so many times and by simplifying the approach to checking for duplicates etc.